requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
pydantic>=2
python-dotenv
redis>=5
httpx
pytest
pytest-asyncio
requests
